% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liliales}
\alias{liliales}
\title{Chloroplast genome sequence similarity matrix for 163 Liliales species}
\format{
## `liliales`
Chloroplast genome sequence similarity matrix for 163 Liliales species
}
\usage{
liliales
}
\description{
Chloroplast genome sequence similarity matrix for 163 Liliales species
}
\keyword{datasets}
