% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixcut.R
\name{matrixcut}
\alias{matrixcut}
\title{Calculates The Component Membership in a Sequence Similarity Matrix at a given Cutoff Value}
\arguments{
\item{mx}{a square sequence similarity matrix}

\item{cut}{a given cutoff value to calculate components for, default value = -1}
}
\value{
The inflection point: cutoff value for the optimal number of clusters

A list of the species together with their cluster membership.
}
\description{
The function takes a square sequence similarity matrix and calculates
those independent components in which the sequence similarity between the members
is greater than a specified value (between 0 and 1). The result provied by the function
is a list of species with their component membership.

Version 0.0.1.
Author: Dr. Matthew Cserhati
Email: csmatyi@protonmail.com
March 19, 2023
}
\examples{
matrixcut(primates,0.8)
matrixcut(primates)

}
\references{
Mullner. <ArXiv:1109.2378>

Cserhati, Carter. (2020, Journal of Creation 34(3):41-50), <https://dl0.creation.com/articles/p137/c13759/j34-3_64-73.pdf>
}
