% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primates}
\alias{primates}
\title{Mitochondrial genome sequence similarity matrix for 31 primate species}
\format{
## `primates`
Mitochondrial genome sequence similarity matrix for 31 primate species
}
\usage{
primates
}
\description{
Mitochondrial genome sequence similarity matrix for 31 primate species
}
\keyword{datasets}
