% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xenarthra}
\alias{xenarthra}
\title{Mitochondrial genome sequence similarity matrix for 37 Xenarthra species}
\format{
## `xenarthra`
Mitochondrial genome sequence similarity matrix for 37 Xenarthra species
}
\usage{
xenarthra
}
\description{
Mitochondrial genome sequence similarity matrix for 37 Xenarthra species
}
\keyword{datasets}
